removed return/exit statements from CG loop for easier OpenACC implementation

+++ openmp/stats.f90
@@ -2,7 +2,7 @@ module stats
 
 implicit none
 
-integer(kind=16) :: flops_diff, flops_bc, flops_blas1
-integer(kind=16) :: iters_cg, iters_newton
+integer(kind=8) :: flops_diff, flops_bc, flops_blas1
+integer(kind=8) :: iters_cg, iters_newton
 
 end module stats
